DOC: refactoring
